+++ src/modules/graphics/opengl.h
@@ -14,6 +14,3 @@
 #define GPU_MESH_FIELDS \
   uint32_t vao; \
   uint32_t ibo;
-
-#define GPU_SHADER_FIELDS \
-  uint32_t program;
